fix(config): cast the security channel retention to int
CI builds .env from .env.example, so LOG_SECURITY_DAYS arrives as the
string '90' and the channel config pin fails; local runs without the
env line got the int default and passed.

+++ config/logging.php
@@ -66,7 +66,7 @@ return [
             'driver' => 'daily',
             'path' => storage_path('logs/security.log'),
             'level' => 'info',
-            'days' => env('LOG_SECURITY_DAYS', 90),
+            'days' => (int) env('LOG_SECURITY_DAYS', 90),
             'replace_placeholders' => true,
         ],
